refactor: useSelect hook file name
