+++ src/components/hooks/index.ts
@@ -4,4 +4,4 @@ export * from './useModifiedPopper';
 export * from './useOnClickOutside';
 export * from './useOnOff';
 export * from './useToggle';
-export * from './useState';
+export * from './useSelect';
